Fix a hack in the echo client.
Previously, the client object was stored as a global variable, which
wasn't properly initialized until the client actually connected to the
server. So if you tried to type something before then, the client would
just crash. Now we use Deferreds to set up the stdio-handler-to-client
link only once the client is actually initialized. The stdio handler
also checks that the client is initialized before sending messages, so
if you try to type something too early then you'll get a graceful error
message, rather than a client crash.