Rework the definitions of the messages.
Separate posArgs into 4 types, for the 4 different types of coordinates:
  - wPosArg for world coordinates
  - bPosArg for build coordinates
  - cPosArg for chunk coordinates
  - gPosArg for graphics coordinates (client only)

Refactor the way the ArgumentSpecifications are described, so that
different argument types that parse the same can be combined (ex:
{w,b,c}PosArgs are all just pairs of ints).

Fix Click being defined in terms of world coordinates (integers) instead
of graphics coordinates (floats), and reintroduce the code we used to
have to parse a pair of floats.

Add a new message GroundInfo, which will eventually be used for the
server to send information to the client about what type of ground is
present in a particular chunk. Not used yet.