Proof of concept for creating our own animation.
Add a new model, gingerbread.egg, which includes an animated
"gingerbread man". For some reason the texturing didn't work on this
one, so it just shows up as a white outline. But the animation does seem
to work in Panda.

Rather than trying to make it actually walk, I just did an animation of
its arms waving back and forth. Therefore, this animation is called
"wave" instead of "walk".

It's janky enough that I'm keeping it out of the master branch, at least
for now.