On order to nonexistent unit, warn and drop it.
Previously, we would sometimes crash (#22) and sometimes let the order
through but then handle it much later when that unit was actually
created (#28).

In addition to fixing those two issues, I've added a check to make sure
that we don't allow any orders in the UnitOrders mapping that are
assigned to nonexistent units, so we can more easily detect if these
sorts of issues come up again.

Fixes #22. Fixes #27. Fixes #28.