Add Entity class to graphics, & AddEntity message.
In GraphicsInterface, generate an AddEntity message whenever we create
an obelisk, in addition to using the old addObelisk approach. This
means that currently, whenever an obelisk is created, we also create a
second "shadow obelisk" which is embedded in the ground at its initial
location and never moves.

Also, add a couple TODOs for things I noticed while implementing this.