Dynamically scale units, as with the ground.
Currently the target size is hardcoded in GraphicsInterface.

I also fix a bug where we calculate the offset for the unit based on the
center of the model before it's scaled, but apply that offset after
scaling. This became noticeable since we now use that logic to position
the pandas, which seem not to be centered at their origin. (Also the
panda is HUGE, so the difference between the center before scaling and
the center after scaling is on the order of th width of the map.)