Reorganize mouse handling.
Panda events are now routed through two new functions:
  - pandaEventMouseDown
  - pandaEventMouseUp
These keep track of some state to determine whether a "click" or a
"drag" has happened, and if so, they delegate to appropriate handlers.

Making use of this new functionality, we can now call createSelectionBox
and removeSelectionBox. Unfortunately, the latter doesn't work, so I've
commented out that call for now.