Proof of concept for 3d->screen->3d roundtrip.
Before drawing the selection box, we now call coordScreenTo3d and then
immediately coord3dToScreen. There's no visible functional change, but
the point is that we are now demonstrating that those functions are
inverses (at least for sane/default camera positions...).

To get this working, I needed to make a fix to coord3dToScreen, since
the old version was bugged.