Don't blindly forward messages from backend to GI.
The client backend now needs to explicitly indicate if a message from
the server will be forwarded to the graphics interface. Those that had
crept in (because we used to just always forward) are no longer
forwarded. This prevents the backend from forwarding both unvalidated
server messages and also validated server messages that just won't be
handled by the graphics interface.

Fixes #40. Fixes #65.