Drop queued GameStateChange msgs to disconnected clients.
There was a bug where if the server had pended GameStateChanges and
wanted to notify a client about them, but the client disconnected before
the server could send the notification, then
ConnectionManager.sendMessage would crash when trying to lookup the
playerId. Instead, now we detect that and just drop the message.

Fixes #76.